Small Matrix: reduce generic kernel source files
